backend/requirements.txt
Django
djangorestframework
django-cors-headers
boto3
pandas
openpyxl
mysqlclient
